+++ src/test/resources/com/atomist/rug/test/gherkin/handler/command/CommandHandlers.ts
@@ -1,7 +1,9 @@
 import {HandleCommand, Instruction, Response, HandlerContext, Plan, Message} from '@atomist/rug/operations/Handlers'
 import {CommandHandler, Secrets, Parameter, Tags, Intent} from '@atomist/rug/operations/Decorators'
 
-@CommandHandler("ReturnsEmptyPlanCommandHandler","Search Youtube for kitty videos and post results to slack")
+import * as node from "./Nodes"
+
+@CommandHandler("ReturnsEmptyPlanCommandHandler","Return empty plan")
 @Tags("kitty", "youtube", "slack")
 @Intent("show me kitties","cats please")
 @Secrets("atomist/user_token", "atomist/showmethemoney")
@@ -18,18 +20,61 @@ class ReturnsEmptyPlanCommandHandler implements HandleCommand {
 export let command1 = new ReturnsEmptyPlanCommandHandler();
 
 
-@CommandHandler("ReturnsOneMessageCommandHandler","Search Youtube for kitty videos and post results to slack")
+@CommandHandler("ReturnsOneMessageCommandHandler","Returns one message")
 @Tags("kitty", "youtube", "slack")
-@Intent("show me kitties","cats please")
 @Secrets("atomist/user_token", "atomist/showmethemoney")
 class ReturnsOneMessageCommandHandler implements HandleCommand {
 
   handle(ctx: HandlerContext) : Plan {
     let result = new Plan()
+    if (ctx.teamId() == null)
+      throw new Error("Cannot get at team id")
+
+    result.add(new Message("woot").withCorrelationId("dude"))
+    console.log(`The constructed plan messages were ${result.messages()},size=${result.messages().length}`)
+    return result;
+  }
+}
+export let command2 = new ReturnsOneMessageCommandHandler();
+
+
+@CommandHandler("RunsPathExpressionCommandHandler","Returns one message")
+@Tags("path_expression")
+class RunsPathExpressionCommandHandler implements HandleCommand {
+
+  handle(ctx: HandlerContext) : Plan {
+    let result = new Plan()
+    const eng = ctx.pathExpressionEngine()
+
+    const findPerson = "/Commit/Person()[@name='Ebony']"
+
+    eng.with<node.Person>(ctx.contextRoot(), findPerson, peep => {
+      throw new Error(`Shouldn't have found a person but found ${peep}`)
+    })
     result.add(new Message("woot").withCorrelationId("dude"))
     console.log(`The constructed plan messages were ${result.messages()},size=${result.messages().length}`)
     return result;
   }
 }
 
-export let command2 = new ReturnsOneMessageCommandHandler();
+export let command3 = new RunsPathExpressionCommandHandler();
+
+
+@CommandHandler("RunsMatchingPathExpressionCommandHandler","Returns one message")
+@Tags("path_expression")
+class RunsMatchingPathExpressionCommandHandler implements HandleCommand {
+
+  handle(ctx: HandlerContext) : Plan {
+    let result = new Plan()
+    const eng = ctx.pathExpressionEngine()
+
+    const findPerson = "/Commit/Person()[@name='Ebony']"
+    eng.with<node.Person>(ctx.contextRoot(), findPerson, peep => {
+      console.log(`Adding message with person name=${peep.name()},obj=${peep}`)
+      result.add(new Message(peep.name()).withCorrelationId("dude"))
+    })
+    console.log(`The constructed plan messages were ${result.messages()},size=${result.messages().length}`)
+    return result;
+  }
+}
+export const command4 = new RunsMatchingPathExpressionCommandHandler();