+++ src/test/resources/com/atomist/rug/test/gherkin/handler/command/NoMessagesFeature1Steps.ts
@@ -1,12 +1,11 @@
 import {Given,When,Then, HandlerScenarioWorld} from "@atomist/rug/test/handler/Core"
-import * as node from "../../../handlers/event/Nodes"
 
 Given("a sleepy country", f => {
+    //console.log("Given invoked for handler")
 })
 When("a visionary leader enters", world => {
-   world.registerHandler("ReturnsEmptyPlanEventHandler2a")
-   let c = new node.Commit().withMadeBy(new node.Person("Ebony"))
-   world.sendEvent(c)
+   let handler = world.commandHandler("ReturnsEmptyPlanCommandHandler")
+   world.invokeHandler(handler, {})
 })
 Then("excitement ensues", world => {
     return world.plan().messages().length == 0