+++ src/test/resources/com/atomist/rug/test/gherkin/handler/command/SingleMessageFeature1StepsWithPathExpression.ts
@@ -1,12 +1,11 @@
 import {Given,When,Then, HandlerScenarioWorld} from "@atomist/rug/test/handler/Core"
-import * as node from "../../../handlers/event/Nodes"
 
 Given("a sleepy country", f => {
 })
 When("a visionary leader enters", world => {
-   world.registerHandler("ReturnsEmptyPlanEventHandler2")
-   world.sendEvent(new node.Commit)
+   let handler = world.commandHandler("RunsPathExpressionCommandHandler")
+   world.invokeHandler(handler, {})
 })
 Then("excitement ensues", world => {
-    return !world.plan()
+    return world.plan().messages().length == 1
 })