+++ src/test/resources/com/atomist/rug/test/gherkin/handler/event/EventHandlers.ts
@@ -50,7 +50,9 @@ class ReturnsEmptyPlanEventHandler3 implements HandleEvent<node.Commit,node.GitH
  
   handle(m: Match<node.Commit, node.GitHubId>) {
     let ghid: node.GitHubId = m.matches()[0]
-    console.log(`Developer's github id=${ghid.id()}`)
+    if (ghid.id() != "gogirl") throw new Error(`Unexpected github id ${ghid.id()}`)
+    if (ghid.address() != "/madeBy/gitHubId") throw new Error(`Unexpected address [${ghid.address()}]`)
+
     return new Plan();
   }
 }